Add ASP.NET Core 3.1 endpoint routing support
Version 1.3.0

This will enable authentication scenarios via endpoint routing.
