+++ build.cmd
@@ -6,4 +6,4 @@ rmdir /s /q "artifacts"
 
 dotnet build --configuration Release
 dotnet test --configuration Release --no-build
-dotnet pack --configuration Release --no-build --output ../../artifacts
+dotnet pack --configuration Release --no-build --output artifacts
