+++ samples/FeatureFlags.AspNetCore/Startup.cs
@@ -1,12 +1,10 @@
-﻿using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using RimDev.AspNetCore.FeatureFlags;
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 
 namespace FeatureFlags.AspNetCore
@@ -31,7 +29,7 @@ namespace FeatureFlags.AspNetCore
             services.AddFeatureFlags(options);
         }
 
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
@@ -39,30 +37,31 @@ namespace FeatureFlags.AspNetCore
             }
 
             app.UseFeatureFlags(options);
-            app.UseFeatureFlagsUI(options);
+            //app.UseFeatureFlagsUI(options);
 
-            app.Map("/test-feature", appBuilder =>
+            app.UseRouting();
+
+            app.UseEndpoints(endpoints =>
             {
-                appBuilder.Run(async context =>
+                endpoints.Map("/test-feature", async context =>
                 {
                     var testFeature = context.RequestServices.GetService<TestFeature>();
 
                     context.Response.ContentType = "text/html";
                     await context.Response.WriteAsync($@"
-                        {testFeature.GetType().Name}: {testFeature.Value}
-                        <br />
-                        <a href=""{options.UiPath}"">View UI</a>");
+                    {testFeature.GetType().Name}: {testFeature.Value}
+                    <br />
+                    <a href=""{options.UiPath}"">View UI</a>");
                 });
-            });
 
-            app.Map("", appBuilder =>
-            {
-                appBuilder.Run(context =>
+                endpoints.Map("", context =>
                 {
                     context.Response.Redirect("/test-feature");
 
                     return Task.CompletedTask;
                 });
+
+                endpoints.MapFeatureFlagsUI(options);
             });
         }
     }
